requirements.txt
langchain
langchain_core
langchain_openai
langserve
streamlit
langchain_community
python-dotenv
langserve
fastapi
uvicorn